requirements.txt
aiohttp==2.2.5
async-timeout==2.0.0
chardet==3.0.4
multidict==3.2.0
pkg-resources==0.0.0
pyserial==3.4
RPi.GPIO==0.6.3
yarl==0.13.0
